obs-ffmpeg: SRT - Remove frame drops, add extradata, some fixes
1)The frame drop mechanism which was a copy paste of what's in rtmp-stream.c does not work well.
There seems indeed to be some interference with srt in-house resend buffer: when packets are lost, upon UNACK
srt stores them in a resend buffer.
As a result whenever, the bandwidth fluctuates strongly, the two mechanisms interfere and this results in a bitrate dropping to 0.
So, for now, we remove all this code for rame drops.
2) The current code for SRT call parts of obs-ffmpeg-output.c (the ffmpeg recording output).
Notably ffmpeg_data_init which is responsabile for creating the AVStreams.
However some unnecessary code paths are called.
WE add a ffmpeg_config.is_encoded_output member to control this.
3) The codec contexts are missing extradata for audio and video; we add them.
4) The audio header was not sent along the video header with the first packets. We've added it.