obs-filters: RTX denoiser initial commit
The RTX denoiser is working.
You'll need to change the model path line 424 to your sdk folder.
To do:
- currently, the nvidia audio FX dll and its deps (about 600 MB !!!) must be moved to obs-plugins folder.
So it'd be cleaner to be able to load the dll from the redist path.
- there's an env var: NVAFX_SDK_DIR which can be queried from registry.
- I got the redist dll loaded and its functions too but wasn't able to have the plugin working.
I must have missed something... So for now, the fx dll is copied manually to obs-plugins.