libobs: Defer reconfiguring encoders to the encode threads
Commit written by r1ch, but I was too lazy to properly rebase so
authorship was not preserved.
(commit 4772a99e3ebaf05)

Fixes a long-standing issue with Dynamic Bitrate where the RTMP output
thread could try to reconfigure an encoder while the encoder is in the
middle of encoding. x264 seems to handle multithreaded calls well, but
NVENC would deadlock in this situation with no error visible to the
user.